requirements.txt
aiogram==3.22.0
pydantic==2.9.2
pydantic-settings==2.6.1
pydantic_core==2.23.4
python-dotenv~=1.1.1